add basic tests measureing a parameter with array validator
